working MDv6 Hall project with no load issue, contains motor on off and direction features

+++ Inc/pmsm_motor_parameters.h
@@ -8,7 +8,7 @@
   ******************************************************************************
   * @attention
   *
-  * <h2><center>&copy; Copyright (c) 2023 STMicroelectronics.
+  * <h2><center>&copy; Copyright (c) 2024 STMicroelectronics.
   * All rights reserved.</center></h2>
   *
   * This software component is licensed by ST under Ultimate Liberty license
@@ -45,7 +45,7 @@
 #define MOTOR_VOLTAGE_CONSTANT  3.8 /*!< Volts RMS ph-ph /kRPM */
 #define NOMINAL_CURRENT_A       9.5
 
-#define ID_DEMAG_A              (int16_t)-9.5 /*!< Demagnetization current */
+#define ID_DEMAG_A              -9.5 /*!< Demagnetization current */
 
 /***************** MOTOR SENSORS PARAMETERS  ******************************/
 /* Motor sensors parameters are always generated but really meaningful only
@@ -68,4 +68,4 @@
                                             revolution */
 
 #endif /* PMSM_MOTOR_PARAMETERS_H */
-/******************* (C) COPYRIGHT 2023 STMicroelectronics *****END OF FILE****/
+/******************* (C) COPYRIGHT 2024 STMicroelectronics *****END OF FILE****/
